+++ tests/helpers.ts
@@ -6,10 +6,25 @@ import {Wallet} from '@ethersproject/wallet';
 import {BigNumber} from '@ethersproject/bignumber';
 import {ContractTransaction} from '@ethersproject/contracts';
 import {MaxUint256} from '@ethersproject/constants';
+import {
+  ZoraModuleManager__factory,
+  WETH__factory,
+  AsksV11__factory,
+  ERC20TransferHelper__factory,
+  ERC721TransferHelper__factory,
+  RoyaltyEngineV1__factory,
+  ZoraProtocolFeeSettings__factory,
+} from '@zoralabs/v3/dist/typechain';
 
 // Type def for catalog protocol contracts
 export type CatalogConfiguredAddresses = {
   cnft: string;
+};
+
+export type ZoraConfiguredAddresses = {
+  asksV11: string;
+  moduleManager: string;
+  weth: string;
 };
 
 export async function setupCatalog(
@@ -26,5 +41,70 @@ export async function setupCatalog(
 
   return {
     cnft: cnftAddress,
+  };
+}
+// Nasty helper function to deploy required zora protocol modules
+export async function setupZora(
+  wallet: Wallet
+): Promise<ZoraConfiguredAddresses> {
+  // setup WETH contract so we have an ERC20 for fees
+  const weth = await (await new WETH__factory(wallet).deploy())._deployed();
+  await weth.deployTransaction.wait();
+  const wethAddress = weth.address;
+
+  // setup the zora module manager contract (moduleManager)
+  const moduleManager = await (
+    await new ZoraModuleManager__factory(wallet).deploy(
+      wallet.address,
+      wethAddress
+    )
+  )._deployed();
+  await moduleManager.deployTransaction.wait();
+  const moduleManagerAddress = moduleManager.address;
+
+  // setup Transfer Helpers
+  const erc20TransferHelper = await (
+    await new ERC20TransferHelper__factory(wallet).deploy(moduleManagerAddress)
+  )._deployed();
+  await erc20TransferHelper.deployTransaction.wait();
+  const erc20TransferHelperAddress = erc20TransferHelper.address;
+
+  const erc721TransferHelper = await (
+    await new ERC721TransferHelper__factory(wallet).deploy(moduleManagerAddress)
+  )._deployed();
+  await erc721TransferHelper.deployTransaction.wait();
+  const erc721TransferHelperAddress = erc721TransferHelper.address;
+
+  // Setup royalty Engine and protocol fee settings
+  const royaltyEngineV1 = await (
+    await new RoyaltyEngineV1__factory(wallet).deploy()
+  )._deployed();
+  await royaltyEngineV1.deployTransaction.wait();
+  const royaltyEngineV1Address = royaltyEngineV1.address;
+
+  const zoraProtocolFeeSettings = await (
+    await new ZoraProtocolFeeSettings__factory(wallet).deploy()
+  )._deployed();
+  await zoraProtocolFeeSettings.deployTransaction.wait();
+  const zoraProtocolFeeSettingsAddress = zoraProtocolFeeSettings.address;
+
+  // setup the asksV11 contract (asksV11)
+  const asksV11 = await (
+    await new AsksV11__factory(wallet).deploy(
+      erc20TransferHelperAddress,
+      erc721TransferHelperAddress,
+      royaltyEngineV1Address,
+      zoraProtocolFeeSettingsAddress,
+      wethAddress
+    )
+  )._deployed();
+
+  await asksV11.deployTransaction.wait();
+  const asksV11Address = asksV11.address;
+
+  return {
+    asksV11: asksV11Address,
+    moduleManager: moduleManagerAddress,
+    weth: wethAddress,
   };
 }